Add files via upload
Added USB conectivity to connect to console and one USB<=>Serial converter. This update was to document the ability to change serial communication paramenters through host. This equipment is complement of PS/2 to MSX key board adapter, so more technical details can be found on that pages. See on the two pdf files more technical details about the design and application.