Update to implement new features and F4 support
Doxygen at working phase. Emulator of MSX Keyboard subsystem, to make tests and performance measurements of the PS/2 to MSX Adatpter. Now implemented the Core M4 suppport.
If USB is enumerated, console is attached to USB device, plus the USART is done as a USB<=>Serial converter. Some interferencies from con to usart was fixed.