Update to implement new features and F4 support
Doxygen in working progress. This firmware is an Emulator of MSX Keyboard subsystem, to make tests and performance measurements of the PS/2 to MSX Adatpter. Now implemented the Core M4 suppport and USB console and UART <=> USB.
If USB is enumerated, console is attached to USB device, and the freeed USART has prized with a USB<=>Serial converter functionality. Some interferencies from con to usart were fixed.