Database compiler
Database compiler upload. Excel file macro enabled PS2toMSX-BR Mapping220901.xlsm
This file is compatible to run with Libre Office.
To run the macro that makes the Database, just click on the MSX Keyboard picture and wait some few seconds to run this VBA, If excel. Libre office will spend a lot of more time to run, as it runs in VBA compatibility node, not an Open Office native program.